spikesorting
Contains pipeline for spike sorting multi-channel data acquired with TDT hardware. Preprocessing steps convert data from TDT native format to -mat files. Files are then converted to binary (-dat), common average referencing is applied, and data is run through kilosort for sorting. In this new version, the pipeline was completed and runs with Kilosort2. This is the current somewhat stable version of the pipeline.